feat: implement stake pools vote tally.
