+++ crates/amaru-stores/src/rocksdb/mod.rs
@@ -19,7 +19,7 @@ use amaru_kernel::{
     StakeCredential, TransactionInput,
 };
 use amaru_ledger::{
-    governance::ratification::ProposalRoots,
+    governance::ratification::ProposalsRoots,
     store::{
         columns as scolumns, Columns, EpochTransitionProgress, HistoricalStores, OpenErrorKind,
         ReadStore, Snapshot, Store, StoreError, TransactionalContext,
@@ -68,8 +68,8 @@ const KEY_PROTOCOL_VERSION: &str = "@protocol-version";
 /// key where is stored the constitutional committee information;
 const KEY_CONSTITUTIONAL_COMMITTEE: &str = "@constitutional-committee";
 
-/// key where are stored the proposal roots;
-const KEY_PROPOSAL_ROOTS: &str = "@proposal-roots";
+/// key where are stored the proposals roots;
+const KEY_PROPOSALS_ROOTS: &str = "@proposals-roots";
 
 /// Name of the directory containing the live ledger stable database.
 const DIR_LIVE_DB: &str = "live";
@@ -342,10 +342,10 @@ macro_rules! impl_ReadStore {
                 get_or_bail(|key| self.db.get(key), &KEY_CONSTITUTIONAL_COMMITTEE)
             }
 
-            fn proposal_roots(
+            fn proposals_roots(
                 &self,
-            ) -> Result<ProposalRoots, StoreError> {
-                get_or_bail(|key| self.db.get(key), &KEY_PROPOSAL_ROOTS)
+            ) -> Result<ProposalsRoots, StoreError> {
+                get_or_bail(|key| self.db.get(key), &KEY_PROPOSALS_ROOTS)
             }
 
             fn pool(&self, pool: &PoolId) -> Result<Option<scolumns::pools::Row>, StoreError> {
@@ -532,9 +532,9 @@ impl TransactionalContext<'_> for RocksDBTransactionalContext<'_> {
         Ok(())
     }
 
-    fn set_proposal_roots(&self, roots: &ProposalRoots) -> Result<(), StoreError> {
+    fn set_proposals_roots(&self, roots: &ProposalsRoots) -> Result<(), StoreError> {
         self.transaction
-            .put(KEY_PROPOSAL_ROOTS, as_value(roots))
+            .put(KEY_PROPOSALS_ROOTS, as_value(roots))
             .map_err(|err| StoreError::Internal(err.into()))?;
         Ok(())
     }
